add logger

+++ RMIAuthServer/src/br/ufrn/imd/pd/AuthService.java
@@ -1,29 +1,30 @@
 package br.ufrn.imd.pd;
 
-import java.rmi.RemoteException; 
-import java.rmi.server.UnicastRemoteObject; 
+import java.rmi.RemoteException;  
 import java.util.List;
 
 import br.ufrn.imd.pd.User; 
 import br.ufrn.imd.pd.DB; 
-
-@SuppressWarnings("serial")
-public class AuthService extends UnicastRemoteObject implements AuthServiceInterface {
+ 
+public class AuthService implements AuthServiceInterface {
 	DB database;
 	User user;
 
 	protected AuthService() throws RemoteException { 
+		System.out.println("Create database");
 		this.database = new DB(); 
 	} 
 
 	@Override
 	public boolean register(String username, String password, List<String> permission) throws RemoteException { 
+		System.out.println("register " +  username + " " +  password + " " + permission.toString());
 		this.database.insert(username, password, permission);
 		return true;
 	}
 
 	@Override
 	public boolean login(String username, String password) throws RemoteException {
+		System.out.println("login " +  username + " " +  password);
 		if (database.validate(username, password)) {
 			user = database.getUser(username);
 			return true;
@@ -34,6 +35,9 @@ public class AuthService extends UnicastRemoteObject implements AuthServiceInter
 
 	@Override
 	public String read(String username) throws RemoteException {
+		System.out.println("read " +  username);
+		System.out.println("user " +  user == null);
+		if (user == null) return "Você precisa estar logado";
 		if (user.getPermission().contains("read")) {
 			return this.database.getUser(username).getPermission().toString();
 		}
@@ -42,6 +46,8 @@ public class AuthService extends UnicastRemoteObject implements AuthServiceInter
 
 	@Override
 	public boolean update(String username, String password, List<String> permission) throws RemoteException {
+		System.out.println("update " +  username + " " +  password + " " + permission.toString());
+		if (user == null) return false;
 		if (user.getPermission().contains("update")) {
 			return this.database.update(username, password, permission);
 		}
@@ -50,6 +56,8 @@ public class AuthService extends UnicastRemoteObject implements AuthServiceInter
 
 	@Override
 	public boolean delete(String username) throws RemoteException {
+		System.out.println("delete " +  username);
+		if (user == null) return false;
 		if (user.getPermission().contains("delete")) {
 			return this.database.delete(username);
 		}
